requirements.txt
kaggle
numpy
pandas
pytest
nltk
sentence-transformers
rake_nltk
matplotlib
wordcloud
seaborn
tqdm
faiss-cpu
streamlit
hf_xet
